+++ issue_priorities.go
@@ -2,8 +2,8 @@ package redmine
 
 import (
 	"encoding/json"
-	"errors"
-	"strings"
+	errors2 "github.com/pkg/errors"
+	"net/http"
 )
 
 type issuePrioritiesResult struct {
@@ -17,23 +17,27 @@ type IssuePriority struct {
 }
 
 func (c *Client) IssuePriorities() ([]IssuePriority, error) {
-	res, err := c.Get(c.endpoint + "/enumerations/issue_priorities.json?" + c.apiKeyParameter() + c.getPaginationClause())
+	url := jsonResourceEndpoint(c.endpoint, "enumerations/issue_priorities")
+	req, err := c.authenticatedGet(url)
 	if err != nil {
-		return nil, err
+		return nil, errors2.Wrap(err, "error while creating GET request for issue priorities")
+	}
+	err = safelySetQueryParameters(req, c.getPaginationClauseParams())
+	if err != nil {
+		return nil, errors2.Wrap(err, "error while adding additional parameters to issue priorities request")
+	}
+	res, err := c.Do(req)
+	if err != nil {
+		return nil, errors2.Wrap(err, "error while reading issue priorities response")
 	}
 	defer res.Body.Close()
 
-	decoder := json.NewDecoder(res.Body)
 	var r issuePrioritiesResult
-	if res.StatusCode != 200 {
-		var er errorsResult
-		err = decoder.Decode(&er)
-		if err == nil {
-			err = errors.New(strings.Join(er.Errors, "\n"))
-		}
-	} else {
-		err = decoder.Decode(&r)
+	if !isHTTPStatusSuccessful(res.StatusCode, []int{http.StatusOK}) {
+		return nil, errors2.Wrapf(decodeHTTPError(res), "issue request returned non-successfully, URL: %s", req.URL.String())
 	}
+
+	err = json.NewDecoder(res.Body).Decode(&r)
 	if err != nil {
 		return nil, err
 	}