+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Add Basic Authentication support for project and issue CRUD requests (#3)
+- Add Basic Authentication support for issue priority read requests (#3)
 - Add option to skip SSL certificate verification (#4)
 
 ### Changed